requirements.txt
face_recognition==1.3.0
scikit-learn==1.4.0
pillow==10.2.0
Jinja2==3.1.3
sqlalchemy==2.0.25
sanic==23.12.1
whoosh==2.7.4
click>=8
